requirements.txt
fastapi
uvicorn
streamlit
requests
python-multipart
pymupdf
sentence-transformers
faiss-cpu
numpy
python-dotenv